backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pg8000==1.30.3
PyJWT==2.8.0
bcrypt==4.0.1
pymupdf==1.23.8
python-docx==1.1.0
google-generativeai==0.3.2
openai==1.6.1
faiss-cpu==1.7.4
sentence-transformers==2.3.1
numpy==1.26.2
requests==2.31.0
gunicorn==21.2.0
flasgger==0.9.7.1
